Call refresh function after creating new project. Refactor. Fixes #2 - Refactored pages - Improved projorg flexibility in creating a new project. - Project list reloads, needing to update the project element in home page.
